moved avataaats into new pages folder
